add client profile api
